+++ hutool-extra/src/main/java/cn/hutool/extra/tokenizer/Word.java
@@ -1,0 +1,17 @@
+package cn.hutool.extra.tokenizer;
+
+/**
+ * 表示分词中的一个词
+ * 
+ * @author looly
+ *
+ */
+public interface Word {
+
+	/**
+	 * 获取单词文本
+	 * 
+	 * @return 单词文本
+	 */
+	String getText();
+}