+++ hutool-extra/src/main/java/cn/hutool/extra/tokenizer/engine/mmseg/MmsegEngine.java
@@ -1,0 +1,51 @@
+package cn.hutool.extra.tokenizer.engine.mmseg;
+
+import java.io.StringReader;
+
+import com.chenlb.mmseg4j.ComplexSeg;
+import com.chenlb.mmseg4j.Dictionary;
+import com.chenlb.mmseg4j.MMSeg;
+import com.huaban.analysis.jieba.JiebaSegmenter.SegMode;
+
+import cn.hutool.core.util.StrUtil;
+import cn.hutool.extra.tokenizer.Engine;
+import cn.hutool.extra.tokenizer.Result;
+
+/**
+ * mmseg4j分词引擎实现<br>
+ * 项目地址：https://github.com/chenlb/mmseg4j-core
+ * 
+ * @author looly
+ *
+ */
+public class MmsegEngine implements Engine {
+
+	private MMSeg mmSeg;
+	
+	/**
+	 * 构造
+	 * 
+	 * @param mode 模式{@link SegMode}
+	 */
+	public MmsegEngine() {
+		this(SegMode.INDEX);
+	}
+	
+	/**
+	 * 构造
+	 * 
+	 * @param mode 模式{@link SegMode}
+	 */
+	public MmsegEngine(SegMode mode) {
+		final Dictionary dict = Dictionary.getInstance();
+		final ComplexSeg seg = new ComplexSeg(dict);
+		this.mmSeg = new MMSeg(new StringReader(""), seg);
+	}
+
+	@Override
+	public Result parse(CharSequence text) {
+		this.mmSeg.reset(StrUtil.getReader(text));
+		return new MmsegResult(this.mmSeg);
+	}
+
+}