+++ hutool-extra/src/main/java/cn/hutool/extra/tokenizer/engine/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * 第三方分词引擎实现
+ * 
+ * @author looly
+ *
+ */
+package cn.hutool.extra.tokenizer.engine;